standardize keeping query string on early questions
